tests for TransactionView
