+++ ethcore/src/views/mod.rs
@@ -1,0 +1,5 @@
+mod transaction;
+mod views;
+
+pub use self::transaction::TransactionView;
+pub use self::views::*;
